+++ src/Supervertaler.Trados/Controls/GroupShareSettingsPanel.cs
@@ -1,8 +1,9 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
+using Supervertaler.Core;
 using Supervertaler.Trados.Core;
 using Supervertaler.Trados.Settings;